I1M.Temas.T1

Description

Documentation

doble :: Num a => a -> a

(doble x) es el doble de x. Por ejemplo,

doble 3         == 6
doble (doble 3) == 12

prop_doble :: (Eq a, Num a) => a -> a -> Bool

prop_doble se verifica si para todo par de números x e y, el doble de x más y es el doble de x más el doble de y. Por ejemplo,

ghci> quickCheck prop_doble
+++ OK, passed 100 tests.

prop_prod_suma :: (Eq a, Num a) => a -> a -> Bool

prop_prod_suma se verifica si para todo par de números, su producto es distinto su suma. Por ejemplo,

ghci> quickCheck prop_prod_suma
*** Failed! Falsifiable (after 1 test):
0
0

prop_prod_suma' :: (Eq a, Num a) => a -> a -> Property

prop_prod_suma' se verifica si el producto de dos números no nulos cualesquiera es distinto de su suma. Por ejemplo, > ghci> quickCheck prop_prod_suma' > +++ OK, passed 100 tests. > ghci> quickCheck prop_prod_suma' > *** Failed! Falsifiable (after 5 tests): > 2 > 2

suma :: (Enum a, Num a) => a -> a

(suma n) es la suma de los n primeros números. Por ejemplo,

suma 4 == 10

sumaLista :: Num a => [a] -> a

(sumaLista xs) es la suma de los elementos de xs. Por ejemplo,

sumaLista [2,3,7] == 12

ordena :: Ord t => [t] -> [t]

(ordena xs) es la lista obtenida ordenando xs mediante el algoritmo de ordenación rápida. Por ejemplo,

ordena [4,6,2,5,3] == [2,3,4,5,6]
ordena "deacb"     == "abcde"